+++ device-bootloaders/src/ServiceProcessorBL2/include/thermal_pwr_mgmt.h
@@ -42,6 +42,18 @@
 /* Number of samples to collect average */
 #define NUM_SAMPLES 100
 
+/*! \struct module_current_t
+    \brief Struct use to hold samples current values
+*/
+struct module_current_t
+{
+    uint8_t minion; /**< Minion Current */
+    uint8_t noc;    /**< NOC Current */
+    uint8_t sram;   /**< SRAM Current */
+    uint8_t pad[5]; /**< Padding for alignment */
+
+} __attribute__((packed));
+
 /*! \fn volatile struct soc_power_reg_t *get_soc_power_reg(void)
     \brief Interface to get the SOC power register
     \param none