+++ device-bootloaders/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added locks while configure/reset SP trace
+- [SW-13253] Added support to read Minion, NOC and SRAM current from PMIC
 ### Changed
 ### Deprecated
 ### Removed